--- classes/game_item.rb
@@ -1,31 +1,0 @@
-require 'date'
-
-class GameItem
-  attr_accessor :archived, :publish_date, :id
-
-  def initialize(publish_date, id = nil)
-    @id = id || Random.rand(1..1000)
-    @publish_date = publish_date
-  end
-
-  def add_genre; end
-
-  def add_author(author)
-    @author = author
-    author.items << self unless author.items.include?(self)
-  end
-
-  def add_label; end
-
-  def add_source; end
-
-  def can_be_archived?
-    current_year = Date.today.year
-    publish_year = Date.parse(@publish_date).year
-    current_year - publish_year > 10
-  end
-
-  def move_to_archive
-    @archived = true if can_be_archived?
-  end
-end
