+++ classes/item.rb
@@ -12,7 +12,10 @@ class Item
     @genre = genre
   end
 
-  def add_author; end
+  def add_author(author)
+    @author = author
+    author.items << self unless author.items.include?(self)
+  end
 
   def add_label(label)
     @label = label
