+++ data/author.json
@@ -3,5 +3,10 @@
     "id": 208,
     "first_name": "greg",
     "last_name": "steve"
+  },
+  {
+    "id": 194,
+    "first_name": "salim",
+    "last_name": "greg"
   }
 ]
