+++ data/games.json
@@ -5,5 +5,12 @@
     "multiplayer": true,
     "last_played_at": "2023/5/4",
     "author": "greg"
+  },
+  {
+    "id": 584,
+    "publish_date": "2023/5/6",
+    "multiplayer": true,
+    "last_played_at": "2023/7/6",
+    "author": "salim"
   }
 ]
