+++ modules/modules.rb
@@ -12,6 +12,11 @@ require_relative 'music/addmusic'
 require_relative 'music/displaymusic'
 require_relative 'music/displaygenre'
 require_relative 'music/storage'
+require_relative 'game/addgame'
+require_relative 'game/displaygame'
+require_relative 'game/displayauthor'
+require_relative 'game/preservegameauthor'
+require_relative 'game/load_game_author'
 
 module Modules
   include AddMovie
@@ -28,6 +33,11 @@ module Modules
   include DisplayMusic
   include DisplayGenres
   include Storage
+  include AddGame
+  include DisplayGame
+  include DisplayAuthor
+  include PreserveGameAuthor
+  include LoadGameAuthor
 
   def save_collections
     save_movies
@@ -35,6 +45,8 @@ module Modules
     save_book
     save_label
     save_music
+    save_games
+    save_author
   end
 
   def load_collections
@@ -44,5 +56,7 @@ module Modules
     load_books
     load_genres
     load_music_albums
+    load_authors
+    load_games
   end
 end
